refactor: extract TransformationType enum from executor
