+++ src/enums/transformation-type.enum.ts
@@ -1,0 +1,5 @@
+export enum TransformationType {
+  PLAIN_TO_CLASS,
+  CLASS_TO_PLAIN,
+  CLASS_TO_CLASS,
+}
